static build
