+++ .gitignore
@@ -13,7 +13,6 @@
 
 # next.js
 /.next/
-/out/
 
 # production
 /build
