+++ frontend/.gitignore
@@ -11,7 +11,6 @@
 
 # next.js
 /.next/
-/out/
 
 # production
 /build
