+++ frontend/package.json
@@ -15,10 +15,11 @@
     "flowbite": "^2.3.0",
     "flowbite-react": "^0.7.5",
     "framer-motion": "^11.0.24",
-    "next": "14.1.4",
+    "next": "^14.1.4",
     "react": "^18",
     "react-dom": "^18",
-    "react-masonry-css": "^1.0.16"
+    "react-masonry-css": "^1.0.16",
+    "serve": "^14.2.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
