+++ frontend/src/app/news/page.js
@@ -1,17 +1,93 @@
-
 "use client";
+import React, { useState, useEffect, useRef } from "react";
+import { Accordion, AccordionItem } from "@nextui-org/react";
 
-import { Button, Card } from "flowbite-react";
+export default function App() {
+  const [news, setNews] = useState([]);
+  const [pageNumber, setPageNumber] = useState(1);
+  const [loading, setLoading] = useState(false);
+  const [hasMore, setHasMore] = useState(true);
 
-export default function News() {
+  const observer = useRef();
+
+  useEffect(() => {
+    const fetchData = async () => {
+      if (!hasMore || loading) return;
+      setLoading(true);
+      try {
+        const response = await fetch(
+          `http://localhost:8000/api/news/?page=${pageNumber}`
+        );
+        const data = await response.json();
+        if (data.detail === "Invalid page.") {
+          setHasMore(false);
+        } else {
+          setNews((prevNews) => [...prevNews, ...data.results]);
+        }
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+      setLoading(false);
+    };
+
+    fetchData();
+  }, [pageNumber]);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const bottom =
+        Math.ceil(window.innerHeight + window.scrollY) >=
+        document.documentElement.scrollHeight;
+
+      if (bottom && !loading) {
+        setPageNumber((prevPageNumber) => prevPageNumber + 1);
+      }
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, [loading]);
+
+  const readable_time = (time_str) => {
+    const date = new Date(time_str);
+    const year = date.getFullYear();
+    const month = (date.getMonth() + 1).toString().padStart(2, "0");
+    const day = date.getDate().toString().padStart(2, "0");
+    const hours = date.getHours().toString().padStart(2, "0");
+    const minutes = date.getMinutes().toString().padStart(2, "0");
+    const readableTime = ` ${hours}:${minutes} on ${day}-${month}-${year}`;
+
+    return readableTime;
+  };
+
+  const get_actual_source = (source_name) => {
+    if (source_name === "CB") return "Cricbuzz";
+    else if (source_name === "HT") return "Hindustan Times";
+    else return "Unknown";
+  };
+
   return (
-    <Card href="#" className="max-w">
-      <h5 className="text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-        Noteworthy technology acquisitions 2021
-      </h5>
-      <p className="font-normal text-gray-700 dark:text-gray-400">
-        Here are the biggest enterprise technology acquisitions of 2021 so far, in reverse chronological order.
-      </p>
-    </Card>
+    <div className="">
+      <Accordion variant="splitted" className="">
+        {news.map((item, index) => (
+          <AccordionItem
+            key={item.id}
+            aria-label={`Accordion ${index + 1}`}
+            title={item.title}
+            subtitle={`Published : ${readable_time(item.updated_at)} | Source : ${get_actual_source(
+              item.source
+            )}`}
+          >
+            <p className="font-normal text-gray-900 dark:text-gray-400">
+              {item.description}
+            </p>
+          </AccordionItem>
+        ))}
+      </Accordion>
+      {loading && <p>Loading...</p>}
+      {!loading && !hasMore && <p>End of News List</p>}
+    </div>
   );
 }
